fix: changed migration so that rao and alpha amounts are BigIntegers

+++ src/patrol/validation/persistence/migrations/versions/616f4febd39c_create_alpha_sell_event_table.py
@@ -19,11 +19,16 @@ depends_on: Union[str, Sequence[str], None] = None
 
 def upgrade() -> None:
     """Upgrade schema."""
-    op.create_table("alpha_sell_event", sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
-                    sa.Column("created_at", sa.DateTime(timezone=True)), sa.Column("block_number", sa.Integer),
-                    sa.Column("event_type", sa.String), sa.Column("coldkey", sa.String), sa.Column("hotkey", sa.String),
-                    sa.Column("rao_amount", sa.Integer), sa.Column("net_uid", sa.Integer), 
-                    sa.Column("alpha_amount", sa.Integer, nullable=True))
+    op.create_table("alpha_sell_event", 
+                    sa.Column("id", sa.Integer, primary_key=True, autoincrement=True),
+                    sa.Column("created_at", sa.DateTime(timezone=True)), 
+                    sa.Column("block_number", sa.Integer),
+                    sa.Column("event_type", sa.String), 
+                    sa.Column("coldkey", sa.String), 
+                    sa.Column("hotkey", sa.String),
+                    sa.Column("rao_amount", sa.BigInteger), 
+                    sa.Column("net_uid", sa.Integer), 
+                    sa.Column("alpha_amount", sa.BigInteger, nullable=True))
